pyproject.toml
[project]
name = "real-time-financial-data-aggregator-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.6.0",
    "django>=5.2.9",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "redis>=7.1.0",
    "requests>=2.32.5",
]
